SinglePage Auth pages e update de cores
