+++ app/tos/page.jsx
@@ -1,0 +1,5 @@
+const Tos = () => {
+	return <div>Tos</div>;
+};
+
+export default Tos;
