+++ components/sign-up-forms/index.js
@@ -1,0 +1,2 @@
+export { default as LoginForm } from './LoginForm';
+export { default as RegisterForm } from './RegisterForm';
